etl/requirements.txt
pymongo
SQLAlchemy
vaderSentiment
psycopg2-binary